# swagger.yaml
openapi: 3.0.0
info:
  version: "1"
  title: "FSArmy - My Blog API"
  description: >-
    'This is the first professional API project we are creating in full-stack-amy.
    This is a simple blog API application where anyone can create a new account,
    login to their system, and create articles that will be visible publicly'
  contact:
    name: Fs Army
    url: https://github.com/tanzimhossain2
    email: support@fsarmy.com
servers:
  - description: Dev
    url: http://localhost:4000/api/v1
...
paths:
...
  "/articles/{id}":
    get:
      tags: [Article]
      description: Fetch a single article
      parameters:
        - $ref: "#/components/parameters/idPathParams"
        - in: query
          name: expand
          description: expand author and comment object
          schema:
            type: string
            example: "author,comment"
      responses:
        "200":
          description: Single Article view
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Article"
                      - type: object
                        properties:
                          author:
                            $ref: "#/components/schemas/AuthotDTO"

                      - type: object
                        properties:
                          comments:
                            $ref: "#/components/schemas/Comment"

                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        example: "/articles/1"
                      author:
                        type: string
                        example: "/articles/1/author"
                      comments:
                        type: string
                        example: "/articles/1/comments"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
    put:
      security:
        - bearerAuth: []
      tags: [Article]
      description: Create or Update entire article
      parameters:
        - $ref: "#/components/parameters/idPathParams"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int32
                  example: 1
                title:
                  type: string
                  example: article title
                body:
                  type: string
                  format: text
                  example: article body
                cover:
                  type: string
                  format: url
                  example: "https://s3.amazon.com/23"
                status:
                  type: string
                  enum: [draft, published]
                  example: draft
              required: [title, body]
      responses:
        "200":
          description: Successfully Updated
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ArticleUpdateDTO"

        "201":
          description: Successfully Created
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ArticleCreateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"

    patch:
      security:
        - bearerAuth: []
      tags: [Article]
      description: Partically update an article
      parameters:
        - $ref: "#/components/parameters/idPathParams"
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: article title
                body:
                  type: string
                  format: text
                  example: article body
                cover:
                  type: string
                  format: url
                  example: "https://s3.amazon.com/23"
                status:
                  type: string
                  enum: [draft, published]
                  example: draft
      responses:
        "200":
          description: Successfully Updated
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ArticleUpdateDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
    delete:
      security:
        - bearerAuth: []
      tags: [Article]
      description: Delete an article by id
      parameters:
        - $ref: "#/components/parameters/idPathParams"
      responses:
        "204":
          description: Article delete succesfully
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Article:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
            authorId:
              type: integer
              format: int32
              example: 1
            title:
              type: string
              example: first Article
            body:
              type: string
              format: text
              example: The body of the first article
            cover:
              type: string
              format: url
              example: https://se.aws.example.com/22a
            status:
              type: string
              enum: [draft, published]
              example: draft
          required: [id, authorId, title, body, status]
        - $ref: "#/components/schemas/Timestamp"
    ArticleCreateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: Article created succesfuly
        data:
          $ref: "#/components/schemas/Article"
        links:
          type: object
          properties:
            self:
              type: string
              format: url
              example: "/article/1"
            author:
              type: string
              format: url
              example: "/article/1/author"
            comments:
              type: string
              format: url
              example: "/article/1/comments"
    ArticleUpdateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: Article updated Succesfull
        data:
          $ref: "#/components/schemas/Article"

        links:
          type: object
          properties:
            self:
              type: string
              example: "/articles/1"
    Comment:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
            authorId:
              type: integer
              format: int32
              example: 1
            articleId:
              type: integer
              format: int32
              example: 1
            body:
              type: string
              format: text
              example: The body of the Comment
            status:
              type: string
              enum: [public, private]
              example: public
          required: [id, authorId, articleId, body, status]
        - $ref: "#/components/schemas/Timestamp"
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthotDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Tanzim
      required: [id, name]
...
  responses:
    "500":
      description: Internal Server Error
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the incovenience. Please tryagain later.
    "400":
      description: Bad Request
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid Email
                    in:
                      type: string
                      enum: [query, body]
                      example: body
    "401":
      description: Unauthorized
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: You don't have the right permission
    "404":
      description: Resource Not Found
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not Found
              message:
                type: string
                example: Requested resource not found
  parameters:
...
    idPathParams:
      in: path
      name: id
      required: true
      description: Article ID
      schema:
        type: string
        example: 65bf96a173da8f3982e763e2